+++ src/terraform/modules/infra-cicd/main.tf
@@ -7,7 +7,7 @@ locals {
   build_name           = local.resource_prefix
   build_namespace      =  "CICD_BUILD"
   deploy_namespace     = "CICD_DEPLOY"
-  repository_name = local.resource_prefix
+  repository_name      = local.resource_prefix
   source_namespace     = "CICD_SOURCE"
 
 }
@@ -15,10 +15,15 @@ locals {
 locals {
   build_sourcecode = {
     type      = "CODECOMMIT"
-    location = "https://git-codecommit.eu-west-2.amazonaws.com/v1/repos/prod-gregnicoluk-www" # ToDo: unhardcode
+    location = "https://git-codecommit.${data.aws_region.current.id}.amazonaws.com/v1/repos/${local.repository_name}"
     buildspec = var.build_buildspec
     git_clone_depth = 1
   }
+  environment_variables = {
+    CLOUDAUTH_BASE_URL         = var.cloudauth_base_url
+    CLOUDAUTH_CLIENT_ID        = var.cloudauth_client_id
+    CLOUDAUTH_CLIENT_SECRET    = var.cloudauth_client_secret
+    CLOUDAUTH_REDIRECT_URI     = var.cloudauth_redirect_uri
+    CLOUDAUTH_SESSION_DURATION = var.cloudauth_session_duration
+  }
 }
-
-# ToDo: Rename module as it's not just "www" anymore, there is "infra" too. Need to move state too