+++ src/terraform/modules/www-site/cloudfront.tf
@@ -15,12 +15,12 @@ resource "aws_cloudfront_origin_access_identity" "www_s3" {
 
 resource "aws_cloudfront_distribution" "www" {
   # Not using aws-terraform-module because blue/green needs a lifecycle ignore on origin_path
-  lifecycle {
-    ignore_changes = [
-      origin  # Gah!, can't narrow this down to ["public_bucket"]["origin_path"]
-                               # https://github.com/hashicorp/terraform/issues/22504
-    ]
-  }
+//  lifecycle {
+//    ignore_changes = [
+//      origin  # Gah!, can't narrow this down to ["public_bucket"]["origin_path"]
+//                               # https://github.com/hashicorp/terraform/issues/22504
+//    ]
+//  }
 
   aliases             = ["${local.subdomain}.${var.domain_name}"]
   comment             = "CloudFront for ${var.resource_name}"
@@ -46,12 +46,34 @@ resource "aws_cloudfront_distribution" "www" {
     }
   }
 
+  ordered_cache_behavior {
+    # Cache behavior with precedence 0
+    path_pattern           = "/secure/*"
+    target_origin_id       = local.cloudfront_s3_origin_name
+    viewer_protocol_policy = "redirect-to-https"
+    allowed_methods = ["GET", "HEAD", "OPTIONS"]
+    cached_methods = ["GET", "HEAD"]
+    compress = true
+    forwarded_values {
+      query_string = false
+      cookies {
+        forward = "none"
+      }
+    }
+    lambda_function_association {
+      event_type = "viewer-request"
+      include_body = false
+      # ToDo: sort out H/C version/alias
+      lambda_arn   = "${module.lambda_cloudfront_auth.this_lambda_function_arn}:11"
+    }
+  }
+
   default_cache_behavior {
     target_origin_id       = local.cloudfront_s3_origin_name
     viewer_protocol_policy = "redirect-to-https"
 
-    allowed_methods = ["GET", "HEAD"]
-    cached_methods  = ["GET", "HEAD"]
+    allowed_methods = ["GET", "HEAD", "OPTIONS"]
+    cached_methods  = ["GET", "HEAD", "OPTIONS"]
     compress        = true
     forwarded_values {
       query_string = true
@@ -76,4 +98,6 @@ resource "aws_cloudfront_distribution" "www" {
     acm_certificate_arn      = module.acm.this_acm_certificate_arn
     ssl_support_method       = "sni-only"
   }
+
+  tags = var.common_tags
 }