+++ src/terraform/modules/www-site/outputs.tf
@@ -15,3 +15,18 @@ output "cloudfront_distribution_id" {
 output "cloudfront_s3_origin" {
   value = local.cloudfront_s3_origin_name
 }
+# -----------------------------------------------------------------------------
+# Outputs: Cognito
+# -----------------------------------------------------------------------------
+
+output "cognito_user_pool_id" {
+  value = aws_cognito_user_pool.user_pool.id
+}
+
+output "cognito_user_pool_arn" {
+  value = aws_cognito_user_pool.user_pool.arn
+}
+
+output "cognito_user_pool_client_id" {
+  value = aws_cognito_user_pool_client.site_client.id
+}