+++ src/terraform/modules/mod-codebuild/CHANGES.md
@@ -1,2 +1,4 @@
 # details of changes from upstream
  - add git_submodules_config.fetch_submodules
+ - add source_version
+ - build environment vars